Add support for sharing other sellers accounts
This is useful for becoming a poshmark ambassador.
I renamed poshmark_email to poshmark_username for cleaniness. This was causing a bug when the script would try to load the users closest, but the query string will fail to load if it contains a email address. And since a username is required now, I thought it'd be best to rename it to something else.
You can now pass in an argument with the '-a' and give an alternative seller account to share